@@ -1,4 +1,4 @@
-// swift-tools-version:5.1
+// swift-tools-version:5.5
 //
 //  Package.swift
 //  Animator
@@ -12,9 +12,9 @@ import PackageDescription
 let package = Package(
     name: "Animator",
     platforms: [
-        .macOS(.v10_12),
-        .iOS(.v10),
-        .tvOS(.v10)
+        .macOS(.v11),
+        .iOS(.v13),
+        .tvOS(.v13)
     ],
     products: [
         .library(
